Post is working

+++ src/components/checklists/Checklist.js
@@ -8,7 +8,7 @@ import "./checklist.css"
 
 export const ShowChecklist = ({ checklists, setChecklists }) => {
     const [currentUser, setCurrentUser] = useState([])
-    
+
     useEffect(() => {
         getCurrentUser().then((data) => setCurrentUser(data))
     }, [])
@@ -24,28 +24,29 @@ export const ShowChecklist = ({ checklists, setChecklists }) => {
 
         <div className="column">
 
-
             {
 
                 checklists.map(
                     (checklist) => {
+                        if (checklist.user.id === currentUser.id)
 
-                        return <div className = "columns is-4-tablet is-3-desktop">
-                        <div className="card" key={`checklist--${checklist.id}`}>
-                            <div className="card-image has-text-centered px-6">
-                                    <figure className="image_checklist">
-                                        <img src={`http://localhost:8000${checklist.image_url}`}
-                                            alt="Placeholder image" />
-                                    </figure>
-                                </div>
+                            return <div className="columns is-4-tablet is-3-desktop">
+                                <div className="card" key={`checklist--${checklist.id}`}>
+                                    <div className="card-image has-text-centered px-6">
+                                        <figure className="image_checklist">
+                                            <img src={`http://localhost:8000${checklist.image_url}`}
+                                                alt="Placeholder image" />
+                                        </figure>
+                                    </div>
 
                                     <div className="card-content">
-                                        <p>{checklist.title}</p>
-                                        <p>{checklist.task}</p>
-                                        <p>Date: {moment.utc(checklist.publication_date).format("MMMM Do YYYY")} </p>
+                                    <p className='title is-5 mt-3 mb-1'>Title of Task: {checklist.title}</p>
+                                    <p className='title is-5 mt-3 mb-1'>Task Description: {checklist.task}</p>
+                                    <p className='title is-5 mt-3 mb-1'>Date: {moment.utc(checklist.publication_date).format("MMMM Do YYYY")} </p>
                                     </div>
-                                    
+
                                     <div className="has-text-centered">
+
                                         <button className="button is-primary is-small" onClick={() => {
                                             deleteChecklist(checklist.id).then((res) => {
                                                 if (res.status === 304) {
@@ -54,15 +55,13 @@ export const ShowChecklist = ({ checklists, setChecklists }) => {
 
                                             }).then(getChecklists)
                                                 .then(setChecklists)
-                                        }}>Finished</button>
+                                        }}>Finished with Task</button>
                                     </div>
 
-                              
 
+
+                                </div>
                             </div>
-                        </div>
-
-                       
 
 
                     }
